+++ src/main/java/frc/robot/subsystems/LimelightSubsystem.java
@@ -84,7 +84,6 @@ public class LimelightSubsystem extends SubsystemBase {
     SmartDashboard.putNumber("LimelightX", x);
     SmartDashboard.putNumber("LimelightY", y);
     SmartDashboard.putNumber("LimelightArea", area);    
-    
   }
 
     /*public double getDistance() {
